+++ Letter Go/LRScoreManager.h
@@ -14,9 +14,9 @@
 /// An array of the words that have bee submitted this round in teh order they were submitted
 @property (readonly) NSMutableArray *submittedWords;
 /// The score required for the palyer to reach the next level
-@property (readonly) int scoreToNextLevel;
+@property (readonly) NSUInteger scoreToNextLevel;
 /// Returns the player's current score
-@property (readonly) int score;
+@property (readonly) NSUInteger score;
 
 ///The shared instance of the score manager
 + (LRScoreManager *)shared;
@@ -26,12 +26,12 @@
  @param wordDict: Has word stored in "word" key and love letter indices scored in "loveLetters" set
  @return The score for the word
  */
-- (int) submitWord:(NSDictionary *)wordDict;
+- (NSUInteger)submitWord:(NSDictionary *)wordDict;
 
 /*! This function calculates the score for a word, including bonuses for length and love letters
  @param wordDict: Has word stored in "word" key and love letter indices scored in "loveLetters" set
  @return The score for the word
  */
-+ (int) scoreForWordWithDict:(NSDictionary *)wordDict;
++ (NSUInteger)scoreForWordWithDict:(NSDictionary *)wordDict;
 
 @end
